Extract string encryptor creational logic so that it can be reused in the Maven plugin.
